Array and hash now use new technique
treesit--simple-indent et. al. use ( anchor . offset ).  I was
adjusting the position of anchor and just letting offset be 0.  This
does not work.  Effectively I was doing (point) + 2.  But if point is
on a single character followed by a new line, then (point) + 2 now is
on a new line but the column is 0.

I now separate all the cases out in the indent rules and have each
rule deal with a special case.  This might actually make the result
easier to tweak and modify.